+++ CHANGELOG.md
@@ -22,6 +22,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
+## [0.8.1] - 2026-01-23
+
+### Added
+
+- Adds a cached todoList to limit API even more
+
+### Fixed
+
+- Fixes debounce feature within TodosViewForm
+- Sets isStillSaving to false after successful fetch
+
+---
+
 ## [0.8.0] - 2026-01-21
 
 ### Added 
@@ -455,6 +468,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 <!-- [Unreleased]: https://github.com/mnichols08/ctd-react-v3-guided-project/compare/v0.0.1...HEAD -->
 
+[0.8.1]: https://github.com/mnichols08/ctd-react-v3-guided-project/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/mnichols08/ctd-react-v3-guided-project/compare/v0.7.3...v0.8.0
 
 [0.7.3]: https://github.com/mnichols08/ctd-react-v3-guided-project/compare/v0.7.2...v0.7.3